Documents/Riemann-main/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
blobfile==2.0.1
bwsi-grader==1.12.0
CacheControl==0.12.11
certifi==2022.12.7
charset-normalizer==3.0.1
cleo==2.0.1
colorama==0.4.6
crashtest==0.4.1
discord==2.1.0
discord.py==2.1.1
distlib==0.3.6
dulwich==0.20.50
emoji==2.2.0
filelock==3.9.0
frozenlist==1.3.3
guess-language==0.2
html5lib==1.1
idna==3.4
importlib-metadata==6.0.0
jaraco.classes==3.2.3
jsonschema==4.17.3
keyring==23.13.1
lockfile==0.12.2
lxml==4.9.2
more-itertools==9.0.0
msgpack==1.0.4
multidict==6.0.4
numpy==1.24.2
openai==0.26.5
packaging==23.0
pexpect==4.8.0
pkginfo==1.9.6
platformdirs==2.6.2
poetry==1.3.2
poetry-core==1.4.0
poetry-plugin-export==1.3.0
ptyprocess==0.7.0
pycryptodomex==3.17
pyrsistent==0.19.3
python-dotenv==1.0.0
pywin32-ctypes==0.2.0
rapidfuzz==2.13.7
regex==2022.10.31
requests==2.28.2
requests-toolbelt==0.10.1
shellingham==1.5.0.post1
six==1.16.0
tiktoken==0.2.0
tomlkit==0.11.6
tqdm==4.64.1
trove-classifiers==2023.2.20
urllib3==1.26.14
virtualenv==20.19.0
webencodings==0.5.1
yarl==1.8.2
zipp==3.14.0
